2-notebooks/requirements.txt
azure-identity
azure-ai-projects
azure-ai-inference
azure-ai-evaluation
azure-ai-contentsafety
azure-monitor-opentelemetry
azure-ai-inference[opentelemetry]
azure-search-documents
azure-core-tracing-opentelemetry
opentelemetry-exporter-otlp
jupyter
notebook
ipykernel
pandas
tabulate
nbformat
asyncio
typing

